Cleanup of all .i files to be numpy >1.7 compliant.

+++ python/python_src/PyDakota/unit/test_monomial.py
@@ -111,7 +111,7 @@ class TestMonomialApproximation(unittest.TestCase):
 
         # TODO replace this switch with a c++ factory
         from PyDakota.univariate_polynomials import LEGENDRE_ORTHOG
-        basis_types = numpy.asarray([LEGENDRE_ORTHOG]*num_vars)
+        basis_types = numpy.asarray([LEGENDRE_ORTHOG]*num_vars,dtype=numpy.int32)
         opts = {'poly_type':poly_type,'basis_types':basis_types}
         approx = polynomial_approximation_factory(var_transform, opts)
         # if poly_type==PCE:
